refactor(tile): one file per question in space/ and physical/
Three concerns were filed by where they were first written rather than by
what they answer:

- `Geometry` (host) and `RuntimeGeometry` (kernel) are twins that change
  together, but sat in `physical/` and `tile/` with the same rationale
  spelled out twice. Merged into `physical/geometry.rs`, said once.
- `Launcher` and `Space::cube_count`/`cube_dim` are launch-side questions
  that lived under `space/partition/`, which is about how a level splits.
  Now `physical/launch.rs`.
- The lane and split shares were split across two files: the enums in
  `distribution.rs` (the split vocabulary a `Cut` states) and the `Space`
  descent deriving them 200 lines into `space/base.rs`. They are only ever
  read together, so they are now `space/share.rs` — which the compiler
  promptly improved, since the descent must reach `partitioner()` through
  the accessor instead of the private field.

space/base.rs 1031 -> 833, distribution.rs 348 -> 236.
No behaviour change.
