+++ crates/cubek-tile/src/space/base.rs
@@ -4,10 +4,7 @@
 use cubecl::prelude::*;
 use cubecl::zspace::SmallVec;
 
-use crate::{
-    Axis, ByAxis, ComputeScope, Distribution, Instruction, LaneShare, LaneWork, Lanes, LevelRole, MAX_AXES,
-    Partitioner, SplitShare, join_lane_share,
-};
+use crate::{Axis, ByAxis, Distribution, Instruction, LevelRole, MAX_AXES, Partitioner};
 
 /// One axis's size.
 /// `Static` is a comptime constant (a tile edge);
@@ -426,204 +423,6 @@ impl Space {
     pub fn walk_invariant(&self, operand: &Space) -> bool {
         self.axes()
             .all(|axis| self.count(axis) == 1 || !operand.contains(axis))
-    }
-
-    /// What the plane's lanes hold of this space's cells: a `Unit` axis the space doesn't span is
-    /// *folded* across the lanes, so each holds only a partial; one it does span is *carried*,
-    /// giving each lane a cell of its own.
-    ///
-    /// Which lanes hold partials of one cell is a question about the lane index's digits, so the
-    /// answer is a bit mask. `Walk::from_counts` decodes a `Unit` axis as
-    /// `UNIT_POS_X / inner_weight % instances`, which for power-of-two counts is a contiguous run
-    /// of bits; the folded axes' runs are exactly the bits a cell's partials differ in. Fold
-    /// everything and that mask is the whole plane ([`LaneShare::Plane`]); fold under a carry and
-    /// it is a [`LaneShare::Group`], whatever order the axes sit in.
-    /// The share a tile ends up with at the leaf: every level's own share joined, the way
-    /// [`MemData::at`](crate::MemData) joins them one at a time on the way down. A block built
-    /// before the walk descends cannot read the stamped value, but it can compute the value that
-    /// stamping would arrive at, because every level is already known here.
-    pub(crate) fn leaf_lane_share(&self) -> LaneShare {
-        let mut share = LaneShare::Whole;
-        let mut level = self.clone();
-        while !level.is_final() {
-            share = join_lane_share(share, level.lane_share());
-            level = level.divide();
-        }
-        share
-    }
-
-    /// What the plane's lanes are to this space's cells, both halves at once. What a drain is
-    /// built from: [`leaf_lane_share`](Self::leaf_lane_share) alone cannot say who writes.
-    pub(crate) fn lanes(&self) -> Lanes {
-        Lanes {
-            share: self.leaf_lane_share(),
-            work: self.lane_work(),
-        }
-    }
-
-    /// Whether anything rides this space's lanes, across every level: [`Own`](LaneWork::Own) if
-    /// some `Unit` axis is dealt out to more than one lane, [`Repeated`](LaneWork::Repeated) if
-    /// none is and the plane's lanes therefore all run the same work.
-    ///
-    /// Read off the axes rather than off [`cube_dim`](Space::cube_dim), which asks the client for
-    /// the hardware `plane_size` and is a launch-side question; this one is comptime, and what a
-    /// drain needs to know before it elects a writer.
-    pub(crate) fn lane_work(&self) -> LaneWork {
-        let mut level = self.clone();
-        while !level.is_final() {
-            let rides = level.partitioner().axes().into_iter().any(|axis| {
-                let Distribution::Spatial {
-                    scope: ComputeScope::Unit,
-                    coverage,
-                    ..
-                } = level.partitioner().distribution(axis)
-                else {
-                    return false;
-                };
-                coverage.instances_const() != Some(1)
-            });
-            if rides {
-                return LaneWork::Own;
-            }
-            level = level.divide();
-        }
-        LaneWork::Repeated
-    }
-
-    /// What one instance of `axes`' operand holds of its cells: [`Partial`](SplitShare::Partial)
-    /// where some `Plane` or `Cube` axis the operand does not span is dealt out across more than
-    /// one instance, so each contracts a slice and none of their results is a whole cell.
-    ///
-    /// [`lane_share`](Self::lane_share)'s counterpart, and asked of the *whole* space rather than
-    /// of the operand's own projection. The projection is what drops the contracted axis, and an
-    /// axis it has dropped has no extent left to divide, so a projected space cannot tell a split
-    /// from a cut whose edge happens to be the whole axis. `Unit` is not among the scopes here
-    /// because a plane's lanes combine in registers, which is
-    /// [`lane_share`](Self::lane_share)'s subject.
-    ///
-    /// Answered conservatively where the instance count is not comptime, which is any
-    /// [`Dynamic`](Extent::Dynamic) extent under [`TilesEach`](Coverage::TilesEach): only the
-    /// shape knows how many tiles it has. Calling that partial costs a fold a one-instance grid
-    /// did not need, while calling it whole loses every partial but one. The same bargain
-    /// [`spans_contracted_at_leaf`](Self::spans_contracted_at_leaf) strikes.
-    pub(crate) fn split_share_of(&self, axes: &[Axis]) -> SplitShare {
-        let mut level = self.clone();
-        while !level.is_final() {
-            let split = level.partitioner().axes().into_iter().any(|axis| {
-                // An axis the operand spans is carried, not split: it gives each instance a cell
-                // of its own rather than a slice of one.
-                if axes.contains(&axis) {
-                    return false;
-                }
-                let dist = level.partitioner().distribution(axis);
-                match dist.scope() {
-                    Some(ComputeScope::Cube(_)) | Some(ComputeScope::Plane) => {}
-                    Some(ComputeScope::Unit) | None => return false,
-                }
-                level.instances_along(axis) != Some(1)
-            });
-            if split {
-                return SplitShare::Partial;
-            }
-            level = level.divide();
-        }
-        SplitShare::Whole
-    }
-
-    /// How many instances `axis` is dealt out to at this level, where that is comptime: the pinned
-    /// count, or the tile grid divided by each instance's share. `None` where the extent is
-    /// [`Dynamic`](Extent::Dynamic) and so the grid is only known at runtime.
-    fn instances_along(&self, axis: Axis) -> Option<usize> {
-        let coverage = self.partitioner.distribution(axis).coverage();
-        match coverage.instances_const() {
-            Some(instances) => Some(instances),
-            // `TilesEach`: the grid decides, and the grid needs the extent.
-            None => match self.extent_raw(axis) {
-                Extent::Static(_) => Some(coverage.instances(self.count(axis))),
-                Extent::Dynamic => None,
-            },
-        }
-    }
-
-    /// The instance-index weight this space's own axis list cannot see: the product of the
-    /// instance counts of the same-scope axes *inside* `axis* that the partitioner distributes and
-    /// this space does not span.
-    ///
-    /// A projected space is the reason this exists. The instance index is a hardware fact and its
-    /// odometer belongs to the *partitioner* — [`Partitioner::axes`] "keeps every axis of the
-    /// operation, so an output space (`{M, N}`) still names its contraction". An operand that does
-    /// not span a contracted axis still runs on an instance whose index encodes it, so it has to
-    /// divide that axis out to find its own digit. Decoding over the spanned axes alone reads every
-    /// omitted inner axis as weight `1`, which aliases the outer digits onto one value — the same
-    /// list [`lane_share`](Self::lane_share) already reads, for the same reason.
-    ///
-    /// Panics where such an axis carries no comptime instance count: there is nothing to decode
-    /// with, and assuming `1` is exactly the silent aliasing above.
-    pub(crate) fn inner_weight_unspanned(&self, axis: Axis) -> usize {
-        if self.partitioner.is_final() {
-            return 1;
-        }
-        let scope = self.partitioner.distribution(axis).scope();
-        self.partitioner
-            .axes()
-            .iter()
-            .skip_while(|&&a| a != axis)
-            .skip(1)
-            .filter(|&&a| !self.contains(a) && self.partitioner.distribution(a).scope() == scope)
-            .map(|&a| {
-                self.partitioner
-                    .distribution(a)
-                    .coverage()
-                    .instances_const()
-                    .unwrap_or_else(|| panic!(
-                        "Space::inner_weight_unspanned: {a:?} is distributed inside {axis:?} at the \
-                         same scope but this space does not span it, and its instance count is not \
-                         comptime, so {axis:?}'s digit of the instance index cannot be decoded"
-                    ))
-            })
-            .product()
-    }
-
-    pub(crate) fn lane_share(&self) -> LaneShare {
-        if self.partitioner.is_final() {
-            return LaneShare::Whole;
-        }
-        // Innermost first, so `weight` is the axis's stride in the lane index as it is reached,
-        // the same least-significant-last ordering `Walk::from_counts` decodes with.
-        let (mut weight, mut fold_mask) = (1usize, 0usize);
-        for axis in self.partitioner.axes().into_iter().rev() {
-            let Distribution::Spatial {
-                scope: ComputeScope::Unit,
-                coverage,
-                ..
-            } = self.partitioner.distribution(axis)
-            else {
-                continue;
-            };
-            // Asserted, not skipped: a `Unit` axis always resolves to `Instances`
-            // (`Distribution::unit` defers through `PlaneLanes`), and passing over one whose
-            // count we could not read would shift every inner axis's bits by its width.
-            let lanes = coverage
-                .instances_const()
-                .expect("Space::lane_share: a Unit axis must carry a const instance count");
-            if lanes == 1 {
-                continue;
-            }
-            assert!(
-                lanes.is_power_of_two(),
-                "Space::lane_share: {axis:?} rides {lanes} lanes, which is not a power of two, so its partials are not a bit range"
-            );
-            if !self.contains(axis) {
-                fold_mask |= (lanes - 1) * weight;
-            }
-            weight *= lanes;
-        }
-        match fold_mask {
-            0 => LaneShare::Whole,
-            // Every lane's bit folded: nothing is carried, so the plane shares the one cell.
-            mask if mask == weight - 1 => LaneShare::Plane,
-            fold_mask => LaneShare::Group { fold_mask },
-        }
     }
 
     /// The axes in this space but not in `output`, i.e. those contracted.
