+++ crates/cubek-tile/src/space/mod.rs
@@ -3,7 +3,9 @@
 mod base;
 mod partition;
 mod region;
+mod share;
 
 pub use base::*;
 pub use partition::*;
 pub use region::*;
+pub use share::*;
