+++ crates/cubek-tile/src/space/partition/mod.rs
@@ -1,8 +1,6 @@
 //! How a level of the space splits, and the [`Walk`] it produces.
 
 mod distribution;
-mod geometry;
-mod launcher;
 mod operand;
 mod partitioner;
 mod tiling;
@@ -10,7 +8,6 @@ mod walk;
 mod walk_order;
 
 pub use distribution::*;
-pub use launcher::*;
 pub use operand::*;
 pub use partitioner::*;
 pub use tiling::*;
